+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/handlers/fieldMovedHandler.es.js
@@ -41,6 +41,7 @@ export default (props, state, event) => {
 
 	if (parentField && parentField.nestedFields.length === 1) {
 		let parentFieldName = parentField ? parentField.fieldName : '';
+
 		do {
 			if (parentField) {
 				parentFieldName = parentField.fieldName;
