+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-builder/src/main/resources/META-INF/resources/js/components/LayoutProvider/util/fieldset.es.js
@@ -1,0 +1,87 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+import * as FormSupport from 'dynamic-data-mapping-form-renderer/js/components/FormRenderer/FormSupport.es';
+import {PagesVisitor} from 'dynamic-data-mapping-form-renderer/js/util/visitors.es';
+
+import {FIELD_TYPE_FIELDSET} from '../../../util/constants.es';
+import {createField} from '../../../util/fieldSupport.es';
+import {updateField} from '../util/settingsContext.es';
+
+const addNestedFields = ({field, indexes, nestedFields, props}) => {
+	let layout = [{rows: field.rows}];
+	const visitor = new PagesVisitor(layout);
+
+	visitor.mapFields((field, fieldIndex, columnIndex, rowIndex, pageIndex) => {
+		if (
+			!nestedFields.some(
+				(nestedField) => nestedField.fieldName === field.fieldName
+			)
+		) {
+			layout = FormSupport.removeFields(
+				layout,
+				pageIndex,
+				rowIndex,
+				columnIndex
+			);
+		}
+	});
+
+	[...nestedFields].reverse().forEach((nestedField) => {
+		layout = FormSupport.addFieldToColumn(
+			layout,
+			indexes.pageIndex,
+			indexes.rowIndex,
+			indexes.columnIndex,
+			nestedField.fieldName
+		);
+	});
+
+	field = updateField(props, field, 'nestedFields', nestedFields);
+
+	const {rows} = layout[indexes.pageIndex];
+
+	return {
+		...updateField(props, field, 'rows', rows),
+		nestedFields,
+		rows,
+	};
+};
+
+export const createFieldSet = (
+	props,
+	event,
+	nestedFields,
+	rows = [{columns: [{fields: [], size: 12}]}]
+) => {
+	const {fieldTypes} = props;
+	const fieldType = fieldTypes.find((fieldType) => {
+		return fieldType.name === FIELD_TYPE_FIELDSET;
+	});
+	const fieldSetField = createField(props, {...event, fieldType});
+
+	return addNestedFields({
+		field: {
+			...fieldSetField,
+			rows,
+		},
+		indexes: {
+			columnIndex: 0,
+			pageIndex: 0,
+			rowIndex: 0,
+		},
+		nestedFields,
+		props,
+	});
+};
